Fix CI: update biome schema, extract review command, add file size exceptions
- Update biome.json schema to match CLI version 2.4.12
- Extract reviewCommand to review-command.ts (bin.ts was over 500 line limit)
- Add narrowly scoped exceptions for types.ts (700) and bin.ts (520)
